REFACTOR: devide textquest.dart into controller and view
