+++ lib/pages/quest/view/component/quest_component.dart
@@ -3,10 +3,11 @@ import 'package:flutter_svg/svg.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:frontend/fastAPI/models/Quest.dart';
 import 'package:frontend/pages/quest/controller/quest_controller.dart';
+import 'package:frontend/pages/questdetails/view/imagequest.dart';
 import 'package:get/get.dart';
 
 import '../../../../common/common.dart';
-import '../../../questdetails/imagequest/view/imagequest.dart';
+import '../../../questdetails/view/textquest.dart';
 
 class QuestComponent extends StatelessWidget {
   final Quest model;
@@ -20,7 +21,7 @@ class QuestComponent extends StatelessWidget {
 
     return InkWell(
         onTap: () async {
-          Get.to(() => UploadImagePage());
+          Get.to(() => UploadTextPage());
         },
         child: SizedBox(
           width: 341,
